backend/pyproject.toml
[project]
name = "panopticon-backend"
version = "0.1.0"
description = "Panopticon multi-agent orchestration backend (LangChain/LangGraph)"
authors = [{ name = "Tarun", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "langchain-core>=0.3,<1.0",
    "langchain-anthropic>=0.3,<1.0",
    "langgraph>=0.4,<1.0",
    "langchain-community>=0.3,<1.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
